# api.yml
openapi: 3.0.3
info:
  title: RSU/ESOP Platform API
  description: |
    API for managing equity compensation (RSUs/stock options) for startups.
    Features include pool management, grant administration, vesting calculations, and employee portal access.
    
    Key platform features:
    - Single ordinary-share class
    - Fixed DECIMAL(12,3) precision for all share & monetary amounts
    - 12-month cliff, 48-month total vesting
    - Roles: Admin, Employee
    - Pool & PPS management with vesting engine
    - Daily vesting batch job
    - Soft delete on all tables
  version: '1.0.0'
  contact:
    name: Platform Support
    email: support@rsu-platform.com
servers:
  - url: 'https://api.rsu-platform.com/api'
    description: Production API server
  - url: 'https://staging.rsu-platform.com/api'
    description: Staging server for testing
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Auth0-issued JWT Access Token.
        Obtained via Auth0 Universal Login flow.
        Contains standard claims plus potential custom claims for role/permissions.
        Example claims expected by backend middleware:
        - sub: Auth0 User ID (used as auth0_user_id)
        - aud: API Identifier (audience)
        - iss: Issuer URL
        - exp: Expiration timestamp
        - iat: Issued at timestamp
        - scope: Permissions (if using RBAC)
        - Custom claims for role/tenant_id (if configured in Auth0 Rules/Actions)
  schemas:
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              example: 'INVALID_GRANT'
            message:
              type: string
              description: Human-readable error message
              example: 'The specified grant does not exist or is inaccessible'
            details:
              type: object
              description: Field-specific validation errors
              example:
                grant_id: 'Invalid UUID format'
                share_amount: 'Must be greater than 0.001'
                status: 'Must be one of: active, inactive'
    Success:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data, schema varies by endpoint
...
    Employee:
      type: object
      required:
        - employee_id
        - tenant_id
        - email
        - first_name
        - last_name
        - status
        - created_at
      properties:
        employee_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        tenant_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        email:
          type: string
          format: email
          example: 'employee@example.com'
        first_name:
          type: string
          example: 'Jane'
        last_name:
          type: string
          example: 'Smith'
        status:
          type: string
          enum: [active, inactive]
          example: 'active'
        created_at:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
    EmployeeCreate:
      type: object
      required:
        - email
        - first_name
        - last_name
      properties:
        email:
          type: string
          format: email
          example: 'employee@example.com'
        first_name:
          type: string
          example: 'Jane'
        last_name:
          type: string
          example: 'Smith'
    EmployeeUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
          example: 'jane.smith@example.com'
        first_name:
          type: string
          example: 'Jane'
        last_name:
          type: string
          example: 'Smith'
        status:
          type: string
          enum: [active, inactive]
          example: 'inactive'
    EmployeeBulkCreate:
      type: object
      required:
        - employees
      properties:
        employees:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeCreate'
          minItems: 1
          maxItems: 100
...
    Grant:
      type: object
      required:
        - grant_id
        - tenant_id
        - employee_id
        - grant_date
        - share_amount
        - vested_amount
        - status
        - created_at
      properties:
        grant_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        tenant_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        employee_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        grant_date:
          type: string
          format: date
          example: '2025-05-01'
        share_amount:
          $ref: '#/components/schemas/DecimalString'
          description: 'Total shares granted (must be > 0)'
        vested_amount:
          $ref: '#/components/schemas/DecimalString'
          description: 'Default is 0.000, increases as shares vest. Atomically synchronized with VestingEvents.'
        status:
          type: string
          enum: [active, inactive]
          example: 'active'
          description: 'Changes from active to inactive upon termination'
        termination_date:
          type: string
          format: date
          nullable: true
          example: null
          description: 'Date of grant termination, if applicable'
        unvested_shares_returned:
          $ref: '#/components/schemas/DecimalString'
          description: 'Calculated as share_amount - vested_amount at termination'
        termination_reason:
          type: string
          nullable: true
          example: null
          description: 'Reason for termination, if applicable'
        terminated_by:
          type: string
          format: uuid
          nullable: true
          example: null
          description: 'User who performed the termination, if applicable'
        version:
          type: integer
          minimum: 0
          example: 0
          description: 'Incremented on each update for optimistic locking'
        created_by:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
          description: 'User who created the grant'
        created_at:
          type: string
          format: date-time
          example: '2025-05-01T10:00:00Z'
        deleted_at:
          type: string
          format: date-time
          nullable: true
          example: null
          description: 'Soft delete timestamp, if applicable'
...
    PaginatedResponse:
      type: object
      required:
        - items
        - pagination
      properties:
        items:
          type: array
          items:
            type: object
        pagination:
          type: object
          required:
            - total_items
            - total_pages
            - current_page
            - limit
          properties:
            total_items:
              type: integer
              example: 45
            total_pages:
              type: integer
              example: 3
            current_page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            next_page:
              type: integer
              nullable: true
              example: 2
            prev_page:
              type: integer
              nullable: true
              example: null
...
    DecimalString:
      type: string
      description: String representation of a DECIMAL(12,3) value.
      pattern: '^-?[0-9]{1,9}(\\.[0-9]{1,3})?$'
      example: '123456.789'
...
  parameters:
...
    EmployeeIdParam:
      name: employee_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the employee
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page for pagination
    StatusParam:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [active, inactive]
      description: Filter by status
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []
paths:
...
  /employees:
    post:
      summary: Create employee
      description: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreate'
      security:
        - bearerAuth: [admin]
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
    
    get:
      summary: List employees
      description: Retrieve paginated list of employees
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/StatusParam'
      security:
        - bearerAuth: [admin]
      responses:
        '200':
          description: Employees retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PaginatedResponse'
                          - type: object
                            properties:
                              items:
                                type: array
                                items:
                                  $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /employees/bulk:
    post:
      summary: Create multiple employees
      description: Create multiple employees in a single request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeBulkCreate'
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Employees created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          created_count:
                            type: integer
                            example: 10
                          employees:
                            type: array
                            items:
                              $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /employees/{employee_id}:
    get:
      summary: Get employee details
      description: Retrieve specific employee details
      parameters:
        - $ref: '#/components/parameters/EmployeeIdParam'
      security:
        - bearerAuth: [admin]
      responses:
        '200':
          description: Employee details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    
    patch:
      summary: Update employee
      description: Update employee details
      parameters:
        - $ref: '#/components/parameters/EmployeeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdate'
      security:
        - bearerAuth: [admin]
      responses:
        '200':
          description: Employee updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    
    delete:
      summary: Delete employee
      description: Soft delete an employee (sets deleted_at timestamp)
      parameters:
        - $ref: '#/components/parameters/EmployeeIdParam'
      security:
        - bearerAuth: [admin]
      responses:
        '200':
          description: Employee deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /employees/{employee_id}/grants-summary:
    get:
      summary: Get employee grants summary
      description: Retrieve summary of all grants for an employee
      parameters:
        - $ref: '#/components/parameters/EmployeeIdParam'
      security:
        - bearerAuth: [admin]
      responses:
        '200':
          description: Grant summary retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          employee_id:
                            type: string
                            format: uuid
                            example: '123e4567-e89b-12d3-a456-426614174000'
                          total_grants:
                            type: integer
                            example: 2
                          total_shares:
                            $ref: '#/components/schemas/DecimalString'
                            example: '200.000'
                          total_vested:
                            $ref: '#/components/schemas/DecimalString'
                            example: '50.000'
                          total_unvested:
                            $ref: '#/components/schemas/DecimalString'
                            example: '150.000'
                          current_pps:
                            $ref: '#/components/schemas/DecimalString'
                            example: '10.500'
                          total_vested_value:
                            $ref: '#/components/schemas/DecimalString'
                            example: '525.000'
                          grants:
                            type: array
                            items:
                              $ref: '#/components/schemas/Grant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...